requirements.txt
requests>=2.28.0
streamlit>=1.28.0
plotly>=5.18.0
python-dotenv>=1.0.0
